% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-docs.R, R/ants-func.R
\name{ants.abp_n4}
\alias{ants.abp_n4}
\title{Truncate outlier intensities and bias correct with the 'N4' algorithm.}
\usage{
ants.abp_n4(
  image,
  intensity_truncation = list(0.025, 0.975, 256L),
  mask = NULL,
  usen3 = FALSE
)
}
\arguments{
\item{image}{'ANTsImage' image to correct and truncate}

\item{intensity_truncation}{3-list quantiles for intensity truncation}

\item{mask}{'ANTsImage' (optional) mask for bias correction}

\item{usen3}{logical, if true, use 'N3' bias correction instead of 'N4' Returns}
}
\description{
Truncate outlier intensities and bias correct with the 'N4' algorithm.
}
\examples{
if( interactive() && ants_available() ) {

image <- ants.image_read(ants.get_ants_data('r16'))
image2 = ants.abp_n4(image)

ants.plot(image)
ants.plot(image2)
}
}
