% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-docs.R, R/ants-func.R
\name{ants.add_noise_to_image}
\alias{ants.add_noise_to_image}
\title{Add noise to an image}
\usage{
ants.add_noise_to_image(image, noise_model, noise_parameters)
}
\arguments{
\item{image}{'ANTsImage' scalar image}

\item{noise_model}{string \code{'additivegaussian'}, \code{'saltandpepper'},
\code{'shot'}, or \code{'speckle'}}

\item{noise_parameters}{\code{\link{tuple}} or array or float, depending on
\code{noise_model}. For \code{'additivegaussian'}, this is (mean,
standard deviation), for \code{'saltandpepper'} (probability, salt-value,
pepper-value), for \code{'shot'} (scale), for \code{'speckle'} (standard
deviation)}
}
\description{
Add noise to an image using additive Guassian, salt-and-pepper, shot, or speckle noise.
}
\examples{

if( interactive() && ants_available() ) {

image = ants.image_read(ants.get_ants_data('r16'))
noise_image1 = ants.add_noise_to_image(
  image, 'additivegaussian', list(0.0, 1.0))
noise_image2 = ants.add_noise_to_image(
  image, 'saltandpepper', list(0.1, 0.0, 100.0))
noise_image3 = ants.add_noise_to_image(image, 'shot', 1.0)
noise_image4 = ants.add_noise_to_image(image, 'speckle', 1.0)

}

}
