% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-docs.R, R/ants-func.R
\name{ants.affine_initializer}
\alias{ants.affine_initializer}
\title{A multi-start optimizer for affine registration}
\usage{
ants.affine_initializer(
  fixed_image,
  moving_image,
  search_factor = 20L,
  radian_fraction = 0.1,
  use_principal_axis = FALSE,
  local_search_iterations = 10L,
  mask = NULL,
  txfn = NULL
)
}
\arguments{
\item{fixed_image}{: ANTsImage:
the fixed reference image moving_image : ANTsImage the moving image to be mapped to the fixed space search_factor : scalar degree of increments on the sphere to search radian_fraction : scalar between zero and one, defines the arc to search over use_principal_axis : boolean boolean to initialize by principal axis local_search_iterations : scalar gradient descent iterations mask : ANTsImage (optional) optional mask to restrict registration txfn : string (optional) filename for the transformation}

\item{moving_image}{: ANTsImage :
the moving image to be mapped to the fixed space search_factor : scalar degree of increments on the sphere to search radian_fraction : scalar between zero and one, defines the arc to search over use_principal_axis : boolean boolean to initialize by principal axis local_search_iterations : scalar gradient descent iterations mask : ANTsImage (optional) optional mask to restrict registration txfn : string (optional) filename for the transformation}

\item{search_factor}{: scalar:
degree of increments on the sphere to search radian_fraction : scalar between zero and one, defines the arc to search over use_principal_axis : boolean boolean to initialize by principal axis local_search_iterations : scalar gradient descent iterations mask : ANTsImage (optional) optional mask to restrict registration txfn : string (optional) filename for the transformation}

\item{radian_fraction}{: scalar:
between zero and one, defines the arc to search over use_principal_axis : boolean boolean to initialize by principal axis local_search_iterations : scalar gradient descent iterations mask : ANTsImage (optional) optional mask to restrict registration txfn : string (optional) filename for the transformation}

\item{use_principal_axis}{: boolean:
boolean to initialize by principal axis local_search_iterations : scalar gradient descent iterations mask : ANTsImage (optional) optional mask to restrict registration txfn : string (optional) filename for the transformation}

\item{local_search_iterations}{: scalar:
gradient descent iterations mask : ANTsImage (optional) optional mask to restrict registration txfn : string (optional) filename for the transformation}

\item{mask}{: ANTsImage (optional):
optional mask to restrict registration txfn : string (optional) filename for the transformation}

\item{txfn}{: string (optional):
filename for the transformation}
}
\description{
Searches over the sphere to find a good initialization for further
registration refinement, if needed. This is a arapper for the ANTs
function antsAffineInitializer. ANTsR function: \code{affineInitializer} Arguments
}
\examples{

if( interactive() && ants_available() ) {

fi <- ants.image_read(ants.get_ants_data('r16'))
mi <- ants.image_read(ants.get_ants_data('r27'))
txfile <- ants.affine_initializer( fi, mi )
tx <- ants.read_transform(txfile)
tx

}

}
