% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-docs.R, R/ants-func.R
\name{ants.allclose}
\alias{ants.allclose}
\title{Check if two images have the same array values}
\usage{
ants.allclose(image1, image2)
}
\arguments{
\item{image1}{image1}

\item{image2}{image2}
}
\description{
Check if two images have the same array values
}
