% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-docs.R, R/ants-func.R
\name{ants.anti_alias}
\alias{ants.anti_alias}
\title{Apply Anti-Alias filter to a binary image}
\usage{
ants.anti_alias(image)
}
\arguments{
\item{image}{: ANTsImage:
binary image to which anti-aliasing will be applied}
}
\description{
Apply Anti-Alias filter to a binary image
}
\examples{

if( interactive() && ants_available() ) {


img <- ants.image_read(ants.get_data('r16'))
mask <- ants.get_mask(img)
mask_aa <- ants.anti_alias(img)
mask_aa


}

}
