% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-docs.R, R/ants-func.R
\name{ants.ANTsImage}
\alias{ants.ANTsImage}
\title{ANTs Class ANTsImage}
\usage{
ants.ANTsImage(
  pixeltype = "float",
  dimension = 3L,
  components = 1L,
  pointer = NULL,
  is_rgb = FALSE,
  label_image = NULL
)
}
\arguments{
\item{pixeltype}{pixeltype}

\item{dimension}{dimension}

\item{components}{components}

\item{pointer}{pointer}

\item{is_rgb}{is_rgb}

\item{label_image}{label_image}
}
\description{
ANTs Class ANTsImage
}
