% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-docs.R, R/ants-func.R
\name{ants.ANTsTransform}
\alias{ants.ANTsTransform}
\title{ANTs Class ANTsTransform}
\usage{
ants.ANTsTransform(
  precision = "float",
  dimension = 3L,
  transform_type = "AffineTransform",
  pointer = NULL
)
}
\arguments{
\item{precision}{precision}

\item{dimension}{dimension}

\item{transform_type}{transform_type}

\item{pointer}{pointer}
}
\description{
ANTs Class ANTsTransform
}
