% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-docs.R, R/ants-func.R,
%   R/ants-func_special.R
\name{ants.apply_ants_transform}
\alias{ants.apply_ants_transform}
\alias{ants.apply_ants_transform_to_image}
\alias{ants.apply_ants_transform_to_point}
\alias{ants.apply_ants_transform_to_vector}
\alias{ants.apply_transforms_to_points}
\title{Apply ANTsTransform to data}
\usage{
ants.apply_ants_transform_to_image(
  transform,
  image,
  reference,
  interpolation = "linear"
)

ants.apply_ants_transform_to_point(transform, point)

ants.apply_ants_transform_to_vector(transform, vector)

ants.apply_transforms_to_points(
  dim,
  points,
  transformlist,
  whichtoinvert = NULL,
  verbose = FALSE
)

ants.apply_ants_transform(
  transform,
  data,
  data_type = "point",
  reference = NULL,
  ...
)
}
\arguments{
\item{transform}{: ANTsTransform:
transform to apply to image}

\item{reference}{: ANTsImage:
target space for transforming image}

\item{interpolation}{: string:
type of interpolation to use}

\item{point}{: list/tuple:
point to which the transform will be applied}

\item{vector}{: list/tuple:
vector to which the transform will be applied}

\item{data}{: ndarray/list/tuple:
data to which transform will be applied}

\item{data_type}{: string:
type of data Options : 'point' 'vector' 'image'}

\item{...}{additional options passed to \code{apply_ants_transform_to_image}}
}
\description{
Apply ANTsTransform to data
}
\examples{

if( interactive() && ants_available() ) {

img <- ants.image_read(ants.get_ants_data("r16"))$clone('float')
tx <- ants.new_ants_transform(dimension=2L)
tx$set_parameters(tuple(0.9,0,0,1.1,10,11))
img2 <- tx$apply_to_image(img, img)

ants.plot(img)
ants.plot(img2)


tx <- ants.new_ants_transform()
params <- tx$parameters
tx$set_parameters(params*2)
pt2 <- tx$apply_to_point(tuple(1,2,3))

# should be (2,4,6)
np <- import("numpy")
np$asarray(pt2)


}

}
