% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-plot.R
\name{ants.plot_directory}
\alias{ants.plot_directory}
\title{Create and save an 'ANTsPy' plot in a directory}
\usage{
ants.plot_directory(
  directory,
  recursive = FALSE,
  regex = "*",
  save_prefix = "",
  save_suffix = "",
  axis = NULL,
  ...
)
}
\arguments{
\item{directory}{string:
directory in which to search for images and plot them}

\item{recursive}{boolean:
If true, this function will search through all directories under the given directory recursively to make plots. If false, this function will only create plots for images in the given directory}

\item{regex}{string:
regular expression used to filter out certain filenames or suffixes}

\item{save_prefix}{string:
sub-string that will be appended to the beginning of all saved plot filenames. Default is to add nothing.}

\item{save_suffix}{string:
sub-string that will be appended to the end of all saved plot filenames. Default is add nothing.}

\item{axis}{whether to plot axis}

\item{...}{additional arguments to pass onto the \code{ants.plot} function. e.g. overlay, alpha, cmap, etc. See \code{ants.plot} for more options.}
}
\description{
Create and save an 'ANTsPy' plot for every image matching a
given regular expression in a directory optionally recursively. This is a
good function for quick visualize exploration of all of images in a directory
}
\examples{
if(interactive() && ants_available() &&
  dir.exists('~/desktop/testdir')) {

  ants.plot_directory(directory='~/desktop/testdir',
                      recursive=FALSE, regex='*')
}

}
