% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-plot.R
\name{ants.plot_grid}
\alias{ants.plot_grid}
\title{Plot a collection of images in an arbitrarily-defined grid}
\usage{
ants.plot_grid(
  images,
  slices = NULL,
  axes = 2L,
  figsize = 1,
  rpad = 0L,
  cpad = 0L,
  vmin = NULL,
  vmax = NULL,
  colorbar = TRUE,
  cmap = "Greys_r",
  title = NULL,
  tfontsize = 20L,
  title_dx = 0L,
  title_dy = 0L,
  rlabels = NULL,
  rfontsize = 14L,
  rfontcolor = "white",
  rfacecolor = "black",
  clabels = NULL,
  cfontsize = 14L,
  cfontcolor = "white",
  cfacecolor = "black",
  filename = NULL,
  dpi = 400L,
  transparent = TRUE,
  ...
)
}
\arguments{
\item{images}{: list of ANTsImage types:
image(s) to plot. if one image, this image will be used for all grid locations. if multiple images, they should be arrange in a list the same shape as the \code{gridsize} argument.}

\item{slices}{: integer or list of integers:
slice indices to plot if one integer, this slice index will be used for all images if multiple integers, they should be arranged in a list the same shape as the \code{gridsize} argument}

\item{axes}{: integer or list of integers:
axis or axes along which to plot image slices if one integer, this axis will be used for all images if multiple integers, they should be arranged in a list the same shape as the \code{gridsize} argument}

\item{figsize, rpad, cpad, vmin, vmax, colorbar, cmap, title, tfontsize, title_dx, title_dy, rlabels, rfontsize, rfontcolor, rfacecolor, clabels, cfontsize, cfontcolor, cfacecolor, filename, dpi, transparent}{parameters passing to other methods}
}
\description{
Plot a collection of images in an arbitrarily-defined grid
}
\examples{

if(interactive() && ants_available()) {

  np <- import("numpy")

  mni1 <- ants.image_read(ants.get_data('mni'))
  mni2 <- mni1$smooth_image(1.)
  mni3 <- mni1$smooth_image(2.)
  mni4 <- mni1$smooth_image(3.)

  images <- list(list(mni1, mni2), list(mni3, mni4))
  slices <- list(list(100L, 100L), list(100L, 100L))
  axes <- list(list(2L, 2L), list(2L, 2L))

  # standard plotting
  ants.plot_grid(images=images, slices=slices, title='2x2 Grid')

  ants.plot_grid(images, slices, cpad=0.02, title='Col Padding')
  ants.plot_grid(images, slices, rpad=0.02, title='Row Padding')
  ants.plot_grid(images, slices, rpad=0.02, cpad=0.02,
                 title='Row and Col Padding')
  # Adding plain row and/or column labels
  ants.plot_grid(images, slices,
                 rlabels=list('Row #1', 'Row #2'))
  ants.plot_grid(images, slices,
                 clabels=list('Col #1', 'Col #2'))
  ants.plot_grid(images, slices,
                 rlabels=list('Row 1', 'Row 2'),
                 clabels=list('Col 1', 'Col 2'))

  images <- list(list(mni1, mni2, mni2),
                 list(mni3, mni4, mni4))
  slices <- list(list(100L, 100L, 100L),
                 list(100L, 100L, 100L))
  axes <- list(list(0L, 1L, 2L),
               list(0L, 1L, 2L))
  ants.plot_grid(images, slices, axes,
                 title='Publication Figures with ANTsPy',
                 tfontsize=20, title_dy=0.03, title_dx=-0.04,
                 rlabels=list('Row 1', 'Row 2'),
                 clabels=list('Col 1', 'Col 2', 'Col 3'),
                 rfontsize=16, cfontsize=16)
}


}
