% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-plot.R
\name{ants.plot_hist}
\alias{ants.plot_hist}
\title{Plot a histogram from an ANTsImage}
\usage{
ants.plot_hist(
  image,
  threshold = 0,
  fit_line = FALSE,
  normfreq = TRUE,
  title = NULL,
  grid = TRUE,
  xlabel = NULL,
  ylabel = NULL,
  facecolor = "green",
  alpha = 0.75
)
}
\arguments{
\item{image}{ANTsImage image from which histogram will be created}

\item{threshold}{threshold}

\item{fit_line}{fit_line}

\item{normfreq}{normfreq}

\item{title}{title}

\item{grid}{grid}

\item{xlabel}{xlabel}

\item{ylabel}{ylabel}

\item{facecolor}{facecolor}

\item{alpha}{alpha}
}
\description{
Plot a histogram from an ANTsImage
}
