% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants-plot.R
\name{ants.plot_ortho}
\alias{ants.plot_ortho}
\alias{ants.plot_ortho_double}
\alias{ants.plot_ortho_stack}
\title{Plot an orthographic view of a 3D image}
\usage{
ants.plot_ortho(
  image,
  overlay = NULL,
  reorient = TRUE,
  blend = FALSE,
  xyz = NULL,
  xyz_lines = TRUE,
  xyz_color = "red",
  xyz_alpha = 0.6,
  xyz_linewidth = 2L,
  xyz_pad = 5L,
  orient_labels = TRUE,
  alpha = 1L,
  cmap = "Greys_r",
  overlay_cmap = "jet",
  overlay_alpha = 0.9,
  black_bg = TRUE,
  bg_thresh_quant = 0.01,
  bg_val_quant = 0.99,
  crop = FALSE,
  scale = FALSE,
  domain_image_map = NULL,
  title = NULL,
  titlefontsize = 24L,
  title_dx = 0L,
  title_dy = 0L,
  text = NULL,
  textfontsize = 24L,
  textfontcolor = "white",
  text_dx = 0L,
  text_dy = 0L,
  filename = NULL,
  dpi = 500L,
  figsize = 1,
  flat = FALSE,
  transparent = TRUE
)

ants.plot_ortho_double(
  image,
  image2,
  overlay = NULL,
  overlay2 = NULL,
  reorient = TRUE,
  xyz = NULL,
  xyz_lines = TRUE,
  xyz_color = "red",
  xyz_alpha = 0.6,
  xyz_linewidth = 2L,
  xyz_pad = 5L,
  cmap = "Greys_r",
  alpha = 1L,
  cmap2 = "Greys_r",
  alpha2 = 1L,
  overlay_cmap = "jet",
  overlay_alpha = 0.9,
  overlay_cmap2 = "jet",
  overlay_alpha2 = 0.9,
  black_bg = TRUE,
  bg_thresh_quant = 0.01,
  bg_val_quant = 0.99,
  crop = FALSE,
  scale = FALSE,
  crop2 = FALSE,
  scale2 = TRUE,
  domain_image_map = NULL,
  title = NULL,
  titlefontsize = 24L,
  title_dx = 0L,
  title_dy = 0L,
  text = NULL,
  textfontsize = 24L,
  textfontcolor = "white",
  text_dx = 0L,
  text_dy = 0L,
  filename = NULL,
  dpi = 500L,
  figsize = 1,
  flat = TRUE,
  transpose = FALSE,
  transparent = TRUE
)

ants.plot_ortho_stack(
  images,
  overlays = NULL,
  reorient = TRUE,
  xyz = NULL,
  xyz_lines = FALSE,
  xyz_color = "red",
  xyz_alpha = 0.6,
  xyz_linewidth = 2L,
  xyz_pad = 5L,
  cmap = "Greys_r",
  alpha = 1L,
  overlay_cmap = "jet",
  overlay_alpha = 0.9,
  black_bg = TRUE,
  bg_thresh_quant = 0.01,
  bg_val_quant = 0.99,
  crop = FALSE,
  scale = FALSE,
  domain_image_map = NULL,
  title = NULL,
  titlefontsize = 24L,
  title_dx = 0L,
  title_dy = 0L,
  text = NULL,
  textfontsize = 24L,
  textfontcolor = "white",
  text_dx = 0L,
  text_dy = 0L,
  filename = NULL,
  dpi = 500L,
  figsize = 1,
  colpad = 0L,
  rowpad = 0L,
  transpose = FALSE,
  transparent = TRUE,
  orient_labels = TRUE
)
}
\arguments{
\item{image}{: ANTsImage image to plot}

\item{overlay}{: ANTsImage image to overlay on base image}

\item{xyz}{: list or list of 3 integers if given, solid lines will be drawn to converge at this coordinate. This is useful for pinpointing a specific location in the image.}

\item{cmap}{: string colormap to use for base image. See matplotlib.}

\item{overlay_cmap}{: string colormap to use for overlay images, if applicable. See matplotlib.}

\item{overlay_alpha}{: float level of transparency for any overlays. Smaller value means the overlay is more transparent. See matplotlib.}

\item{black_bg}{: boolean if TRUE, the background of the image(s) will be black. if FALSE, the background of the image(s) will be determined by the values \code{bg_thresh_quant} and \code{bg_val_quant}.}

\item{bg_thresh_quant}{: float if white_bg=TRUE, the background will be determined by thresholding the image at the \code{bg_thresh} quantile value and setting the background intensity to the \code{bg_val} quantile value. This value should be in \link{0, 1} - somewhere around 0.01 is recommended. - equal to 1 will threshold the entire image - equal to 0 will threshold none of the image}

\item{bg_val_quant}{: float if white_bg=TRUE, the background will be determined by thresholding the image at the \code{bg_thresh} quantile value and setting the background intensity to the \code{bg_val} quantile value. This value should be in \link{0, 1} - equal to 1 is pure white - equal to 0 is pure black - somewhere in between is gray}

\item{crop}{: boolean if true, the image(s) will be cropped to their bounding boxes, resulting in a potentially smaller image size. if false, the image(s) will not be cropped}

\item{scale}{: boolean or 2-list if true, nothing will happen to intensities of image(s) and overlay(s) if false, dynamic range will be maximized when visualizing overlays if 2-list, the image will be dynamically scaled between these quantiles}

\item{domain_image_map}{: ANTsImage this input ANTsImage or list of ANTsImage types contains a reference image \code{domain_image} and optional reference mapping named \code{domainMap}. If supplied, the image(s) to be plotted will be mapped to the domain image space before plotting - useful for non-standard image orientations.}

\item{title}{: string add a title to the plot}

\item{filename}{: string if given, the resulting image will be saved to this file}

\item{dpi}{: integer determines resolution of image if saved to file. Higher values result in higher resolution images, but at a cost of having a larger file size}

\item{flat}{: boolean if true, the ortho image will be plot in one row if false, the ortho image will be a 2x2 grid with the bottom left corner blank}

\item{slices}{: list or list of 3 integers slice indices along each axis to plot. This can be absolute array indices (e.g. (80,100,120)), or this can be relative array indices (e.g. (0.4,0.5,0.6)). The default is to take the middle slice along each axis.}

\item{axis}{: integer which axis to plot along if image is 3D}
}
\description{
Use mask_image and/or threshold_image to preprocess images to be be overlaid and display the overlays in a given range.
See the wiki examples. TODO: - add colorbar option ANTsR function: N/A Arguments ---------
}
\examples{

if(interactive() && ants_available()) {

  mni = ants.image_read(ants.get_data('mni'))
  ants.plot_ortho(mni, xyz=list(100L,100L,100L))

  mni2 = mni$threshold_image(7000, mni$max())
  ants.plot_ortho(mni, overlay=mni2)
  ants.plot_ortho(mni, overlay=mni2, flat=TRUE)

  ants.plot_ortho(mni, overlay=mni2, xyz=list(110L,110L,110L),
                  xyz_lines=FALSE, text='Lines Turned Off', textfontsize=12L)

  ants.plot_ortho(mni, mni2, xyz=list(120L,100L,100L),
                  text=' Example Ortho Text', textfontsize=12L,
                  title='Example Ortho Title', titlefontsize=12L)

}


if(interactive() && ants_available()) {

  mni <- ants.image_read(ants.get_data('mni'))
  ch2 <- ants.image_read(ants.get_data('ch2'))
  ants.plot_ortho_double(mni, ch2)

}


if(interactive() && ants_available()) {

  mni = ants.image_read(ants.get_data('mni'))
  ch2 = ants.image_read(ants.get_data('ch2'))

  ants.plot_ortho_stack( list(mni,mni,mni) )

}

}
