% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ants}
\alias{ants}
\alias{load_ants}
\title{Get 'ANTs' module}
\usage{
ants

load_ants(force = FALSE)
}
\arguments{
\item{force}{whether to force reloading \code{ants} module; default is false}
}
\value{
A 'Python' module
}
\description{
Get 'ANTs' module
}
