% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ants_available}
\alias{ants_available}
\title{Check if 'ANTs' is available}
\usage{
ants_available(module = c("ants", "antspynet"))
}
\arguments{
\item{module}{either \code{'ants'} or \code{'antspynet'}; default is
\code{'ants'}}
}
\value{
Logical, whether \code{'ANTs'} or \code{'ANTsPyNet'} is available
}
\description{
Check if 'ANTs' is available
}
\seealso{
\code{\link{install_ants}}
}
