% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{antspynet}
\alias{antspynet}
\alias{load_antspynet}
\title{Get \code{'ANTsPyNet'} module}
\usage{
antspynet

load_antspynet(force = FALSE, error_if_missing = TRUE)
}
\arguments{
\item{force}{whether to force reloading \code{antspynet} module; default is false}

\item{error_if_missing}{whether to raise errors when the module is unable to
load; default is true.}
}
\value{
A 'Python' module if successfully loaded. If \code{error_if_missing}
is set to false and module is unable to load, return \code{NULL}
}
\description{
Get \code{'ANTsPyNet'} module
}
\seealso{
\code{\link{ants}}
}
