% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_extraction.R
\name{antspynet_brain_extraction}
\alias{antspynet_brain_extraction}
\title{Extract brain and strip skull}
\usage{
antspynet_brain_extraction(
  x,
  modality = c("t1", "t1nobrainer", "t1combined", "flair", "t2", "t2star", "bold", "fa",
    "t1t2infant", "t1infant", "t2infant"),
  verbose = FALSE
)
}
\arguments{
\item{x}{input image or image path}

\item{modality}{modality type}

\item{verbose}{whether to print out process to the screen}
}
\value{
Brain mask image
}
\description{
Print \code{antspynet$brain_extraction} to see the original documentation.
}
