% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{halpern_preprocess}
\alias{halpern_preprocess}
\alias{halpern_register_ct_mri}
\alias{halpern_register_template_mri}
\alias{halpern_apply_transform_template_mri}
\title{'ANTs' functions for 'Halpern' lab}
\usage{
halpern_register_ct_mri(
  fixed,
  moving,
  outprefix,
  fixed_is_ct = TRUE,
  verbose = TRUE
)

halpern_register_template_mri(
  fixed,
  moving,
  outprefix,
  mask = NULL,
  verbose = TRUE
)

halpern_apply_transform_template_mri(roi_folder, outprefix, verbose = TRUE)
}
\arguments{
\item{fixed}{fixed image as template}

\item{moving}{moving image that is to be registered into \code{fixed}}

\item{outprefix}{output prefix, needs to be absolute path prefix}

\item{fixed_is_ct}{whether \code{fixed} is 'CT'}

\item{verbose}{whether to verbose the progress; default is true}

\item{mask}{mask file for template (skull-stripped)}

\item{roi_folder}{template 'ROI' or atlas folder in which the image atlases
or masks will be transformed into subject's native brain}
}
\value{
A list of result configurations
}
\description{
'ANTs' functions for 'Halpern' lab
}
