% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{py}
\alias{py}
\title{Get 'Python' main process environment}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
py
}
\value{
The 'Python' main process as a module
}
\description{
Get 'Python' main process environment
}
\keyword{datasets}
