% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{py_list}
\alias{py_list}
\title{List in 'Python'}
\usage{
py_list(..., convert = FALSE)
}
\arguments{
\item{...}{passing to \code{list} ('Python')}

\item{convert}{whether to convert the results back into R; default is no}
}
\value{
List instance, or an R vector if converted
}
\description{
List in 'Python'
}
\examples{


if(interactive() && ants_available()) {

  py_list(list(1,2,3))
  py_list(c(1,2,3))

  py_list(array(1:9, c(3,3)))
  py_list(list(list(1:3), letters[1:3]))

}
}
