% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{py_slice}
\alias{py_slice}
\title{Slice index in 'Python' arrays}
\usage{
py_slice(...)
}
\arguments{
\item{...}{passing to \code{slice} ('Python')}
}
\value{
Index slice instance
}
\description{
Slice index in 'Python' arrays
}
\examples{


if(interactive() && ants_available()) {

  x <- np_array(array(seq(20), c(4, 5)))

  # equivalent to x[::2]
  x[py_slice(NULL, NULL, 2L)]

}
}
