% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tuple}
\alias{py_help}
\alias{np_array}
\alias{import}
\alias{repl_python}
\alias{run_script}
\alias{r_to_py}
\alias{py_to_r}
\alias{py_to_r_wrapper}
\alias{py_str}
\alias{py_run_string}
\alias{py_bool}
\alias{py_dict}
\alias{py_call}
\alias{py_del_attr}
\alias{py_del_item}
\alias{py_eval}
\alias{py_get_attr}
\alias{py_set_attr}
\alias{py_get_item}
\alias{py_set_item}
\alias{py_len}
\alias{py_none}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate:import]{import()}}, \code{\link[reticulate:np_array]{np_array()}}, \code{\link[reticulate:np_array]{np_array()}}, \code{\link[reticulate:py_bool]{py_bool()}}, \code{\link[reticulate:py_call]{py_call()}}, \code{\link[reticulate:py_del_attr]{py_del_attr()}}, \code{\link[reticulate:py_del_item]{py_del_item()}}, \code{\link[reticulate:py_dict]{py_dict()}}, \code{\link[reticulate:py_eval]{py_eval()}}, \code{\link[reticulate:py_get_attr]{py_get_attr()}}, \code{\link[reticulate:py_get_item]{py_get_item()}}, \code{\link[reticulate:py_help]{py_help()}}, \code{\link[reticulate:py_len]{py_len()}}, \code{\link[reticulate:py_none]{py_none()}}, \code{\link[reticulate:py_run_string]{py_run_string()}}, \code{\link[reticulate:py_set_attr]{py_set_attr()}}, \code{\link[reticulate:py_set_item]{py_set_item()}}, \code{\link[reticulate:py_str]{py_str()}}, \code{\link[reticulate:py_to_r]{py_to_r()}}, \code{\link[reticulate:py_to_r_wrapper]{py_to_r_wrapper()}}, \code{\link[reticulate:r_to_py]{r_to_py()}}, \code{\link[reticulate:r_to_py]{r_to_py()}}, \code{\link[reticulate:tuple]{tuple()}}}

  \item{rpymat}{\code{\link[rpymat:repl_python]{repl_python()}}, \code{\link[rpymat:run_script]{run_script()}}}
}}

