% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{t1_preprocess}
\alias{t1_preprocess}
\title{Process 'T1' image}
\usage{
t1_preprocess(
  t1_path,
  templates = "mni_icbm152_nlin_asym_09a",
  work_path = ".",
  verbose = TRUE
)
}
\arguments{
\item{t1_path}{path to a 'T1' image}

\item{templates}{template to use; default is \code{'mni_icbm152_nlin_asym_09a'},}

\item{work_path}{working path, must be a directory}

\item{verbose}{whether to verbose the progress}
}
\value{
Nothing will be returned. Please check \code{work_path} for results.
}
\description{
Process 'MRI' and align with template brains
}
