% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ants_available}
\alias{ants_available}
\title{Check if 'ANTs' is available}
\usage{
ants_available()
}
\value{
Logical, whether 'ANTs' is available
}
\description{
Check if 'ANTs' is available
}
\seealso{
\code{\link{install_ants}}
}
