% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{install_ants}
\alias{install_ants}
\title{Install 'ANTs' via 'ANTsPy'}
\usage{
install_ants(python_ver = "3.11", verbose = TRUE)
}
\arguments{
\item{python_ver}{'Python' version, see \code{\link[rpymat]{configure_conda}};
default is \code{"3.11"} since \code{'ANTsPy'} is compiled for all platforms
under this version}

\item{verbose}{whether to print the installation messages}
}
\value{
This function returns nothing.
}
\description{
Install 'ANTs' via 'ANTsPy'
}
